Fix compile warnings
