+++ examples/rest-server/rest-endpoints.c
@@ -1,6 +1,8 @@
 
 #include "restserver.h"
 
+#include <string.h>
+
 
 static json_t * endpoint_to_json(lwm2m_client_t *client)
 {
