+++ examples/rest-server/rest-utils.c
@@ -1,0 +1,22 @@
+
+#include "rest-utils.h"
+
+#include "restserver.h"
+
+
+int coap_to_http_status(int status)
+{
+    switch (status)
+    {
+    case COAP_204_CHANGED:
+    case COAP_205_CONTENT:
+        return HTTP_200_OK;
+
+    case COAP_404_NOT_FOUND:
+        return HTTP_404_NOT_FOUND;
+
+    default:
+        return -(((status >> 5) & 0x7) * 100 +  (status & 0x1F));
+    }
+}
+
