+++ examples/rest-server/rest-utils.h
@@ -1,0 +1,9 @@
+
+#ifndef REST_UTILS_H
+#define REST_UTILS_H
+
+
+int coap_to_http_status(int status);
+
+#endif // REST_UTILS_H
+
